move image formats into own package
