+++ chunky/src/java/se/llbit/imageformats/png/ITXT.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Chunky.  If not, see <http://www.gnu.org/licenses/>.
  */
-package se.llbit.png;
+package se.llbit.imageformats.png;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
